+++ src/ros/rover/drive/drive_controllers/nova_drive_controller_base/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.16)
-project(drive_controller_base)
+project(nova_drive_controller_base)
 
 # Default to C++17
 if(NOT CMAKE_CXX_STANDARD)
@@ -36,8 +36,8 @@ endforeach()
 
 find_package(generate_parameter_library REQUIRED)
 
-generate_parameter_library(drive_controller_base_parameters
-  src/drive_controller_base.yaml
+generate_parameter_library(nova_drive_controller_base_parameters
+  src/nova_drive_controller_base.yaml
 )
 
 add_library(${PROJECT_NAME} SHARED
@@ -48,27 +48,27 @@ target_include_directories(${PROJECT_NAME}
   $<INSTALL_INTERFACE:include>)
 ament_target_dependencies(${PROJECT_NAME} ${THIS_PACKAGE_INCLUDE_DEPENDS})
 target_link_libraries(${PROJECT_NAME}
-  drive_controller_base_parameters
+  nova_drive_controller_base_parameters
 )
 # Causes the visibility macros to use dllexport rather than dllimport,
 # which is appropriate when building the dll but not consuming it.
-target_compile_definitions(${PROJECT_NAME} PRIVATE "drive_controller_base_BUILDING_DLL")
-pluginlib_export_plugin_description_file(controller_interface drive_controller_base.xml)
+target_compile_definitions(${PROJECT_NAME} PRIVATE "nova_drive_controller_base_BUILDING_DLL")
+pluginlib_export_plugin_description_file(controller_interface nova_drive_controller_base.xml)
 
 
 install(
   DIRECTORY include/
-  DESTINATION include/drive_controller_base
+  DESTINATION include/nova_drive_controller_base
 )
 
-install(TARGETS drive_controller_base drive_controller_base_parameters
-  EXPORT export_drive_controller_base
+install(TARGETS nova_drive_controller_base nova_drive_controller_base_parameters
+  EXPORT export_nova_drive_controller_base
   RUNTIME DESTINATION bin
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
 )
 
-ament_export_targets(export_drive_controller_base HAS_LIBRARY_TARGET)
+ament_export_targets(export_nova_drive_controller_base HAS_LIBRARY_TARGET)
 
 ament_export_dependencies(
   ${THIS_PACKAGE_INCLUDE_DEPENDS}
