+++ src/ros/rover/drive/drive_controllers/nova_drive_controller_base/nova_drive_controller_base.xml
@@ -1,0 +1,5 @@
+<library path="nova_drive_controller_base">
+  <class name="nova_drive_controller_base/NovaDriveControllerBase" type="nova_drive_controller_base::NovaDriveControllerBase" base_class_type="controller_interface::ControllerInterface">
+  <description>Abstract base class for writing drive controllers.</description>
+  </class>
+</library>
